Cleanup and codecov
- rename df to less ambiguous frequency_increment
- add solve_single_time_window to helper functions
- add a test for solve_single_time_window
- factor z-time-series calculation out of coherence into spectral features
